+++ tests/backends/plotly/plotly_canvas_test.py
@@ -87,3 +87,9 @@ def test_save_to_disk(ext):
         fname = os.path.join(path, f'plopp_fig.{ext}')
         canvas.save(filename=fname)
         assert os.path.isfile(fname)
+
+
+def test_save_to_disk_with_bad_extension_raises():
+    canvas = Canvas()
+    with pytest.raises(ValueError):
+        canvas.save(filename='plopp_fig.txt')
