+++ tests/backends/pythreejs/pythreejs_figure_test.py
@@ -5,6 +5,8 @@ from plopp import input_node
 from plopp.data.testing import scatter
 from plopp.backends.pythreejs.figure import Figure
 from plopp.graphics.figscatter3d import FigScatter3d
+import tempfile
+import os
 
 
 def test_log_norm_3d_toolbar_button():
@@ -27,6 +29,6 @@ def test_save_to_html():
                  z='z',
                  norm='log')
     with tempfile.TemporaryDirectory() as path:
-        fname = os.path.join(path, f'plopp_fig.html')
+        fname = os.path.join(path, 'plopp_fig.html')
         fig.save(filename=fname)
         assert os.path.isfile(fname)
